build: migrate to the sborka build conventions
There were no conventions here to replace - the coordinates, the version and
the ktlint wiring sat in allprojects and subprojects blocks in the root, and
publishing.gradle.kts was applied by path into two modules. All of it comes
from ru.workinprogress.sborka 0.1.0.13 now.

The repository had no gradle.properties at all, so it built with no build
cache, no parallel execution and no configuration cache - three defaults
nobody had chosen. Caching and parallel are on. The configuration cache is
not, and that is a finding rather than an omission: three tasks here capture
Gradle script object references, which it cannot serialise. Moving their
bodies into classes with declared inputs is a change to this repository's
test harness rather than to how it is built, and it has an issue of its own.

allWarningsAsErrors, which the conventions turn on and this repository did
not have, found eight things the compiler had been asking about:

- seven opt-ins. The code uses ExperimentalCoroutinesApi and
  ExperimentalSerializationApi and never said so. Said once, in
  compilerOptions, where it is visible: this library depends on APIs its
  authors may change.
- "Expression is unused" in the spin-wait of Mongkn.kt - `else -> { Unit }`
  is now an empty branch with the comment that explains it.
- "Redundant call of conversion method" - bytes.toULong() on a size_t that
  is already a ULong on this platform.

The POM gains the repository url and the scm block, which it did not carry;
the licence stays Apache-2.0 through sborka.licence. The publish credentials
do not move - CI passes the same MONGKN_REPO_USER and MONGKN_REPO_SECRET
under the names the convention reads.

publish.yml had no Java setup at all and ran on whatever JDK the runner image
carried. It worked, and that is the objection: the toolchain is a decision.

build.yml keeps its automatic triggers switched off. That is a written
decision about the sharded contour from M-66 never having run in CI, not an
oversight, and it is not this pull request's to reverse.

Verified locally as far as this host allows - compile, ktlint, and a local
publication whose four modules carry the right group, licence and file names.
The tests need a running mongod; the build workflow is dispatched on this
branch for them.

+++ mongkn-core/src/nativeTest/kotlin/ru/workinprogress/mongkn/bson/BsonAllocations.kt
@@ -112,7 +112,9 @@ private val countingRealloc =
         val result = realloc(mem, bytes)
         when {
             mem == null && result != null -> liveAllocations.fetchAndAdd(1)
-            mem != null && bytes.toULong() == 0uL -> liveAllocations.fetchAndAdd(-1)
+
+            // `size_t` уже ULong на этой платформе — `toULong()` здесь ничего не делал.
+            mem != null && bytes == 0uL -> liveAllocations.fetchAndAdd(-1)
         }
         result
     }
